requirements.txt
matplotlib==2.2.2
numpy==1.16.1
pandas==0.25.0
pyodbc==4.0.23
pytorch-pretrained-bert==0.6.2
PyYAML==3.12
scikit-learn==0.21.2
seaborn==0.8.1
torch==1.3.0
torchvision==0.4.1
psycopg2-binary==2.8.3
nltk==3.3
spacy==2.2.1
scispacy==0.2.2
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.2.4/en_core_sci_md-0.2.4.tar.gz
